+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED
+
+- Added translation examples to the /translations API endpoint (#10)
+
 ## [2.0.0] - 2021-04-29
 
 - The first release of the Python version of the project.